# Conflicts:
#	model/entity/thumbnail.php

+++ model/entity/attachment.php
@@ -1,8 +1,8 @@
 <?php
 /**
- * Nooku Framework - http://nooku.org/framework
+ * Joomlatools Framework - https://www.joomlatools.com/developer/framework/
  *
- * @copyright	Copyright (C) 2011 - 2015 Johan Janssens and Timble CVBA. (http://www.timble.net)
+ * @copyright	Copyright (C) 2011 Johan Janssens and Timble CVBA. (http://www.timble.net)
  * @license		GNU GPLv3 <http://www.gnu.org/licenses/gpl.html>
  * @link		http://github.com/joomlatools/joomlatools-framework-files for the canonical source repository
  */
@@ -25,25 +25,19 @@ class ComFilesModelEntityAttachment extends KModelEntityRow
         return $this->getObject('com:files.model.files')
                     ->container($this->container_slug)
                     ->name($this->name)
+                    ->thumbnails(true)
                     ->fetch()
                     ->getIterator()
                     ->current();
     }
 
-    /**
-     * Overridden for deleting the attached file.
-     */
     public function delete()
     {
-        $result = parent::delete();
-
-        if ($result)
+        if ($result = parent::delete())
         {
             $file = $this->file;
 
-            if (!$file->isNew()) {
-                $file->delete();
-            }
+            if (!$file->isNew()) $file->delete();
         }
 
         return $result;
@@ -55,17 +49,8 @@ class ComFilesModelEntityAttachment extends KModelEntityRow
 
         $file = $this->file;
 
-        if (!$file->isNew())
-        {
-            $data['file'] = $file->toArray();
-
-            if ($file->isThumbnail())
-            {
-                $thumbnail = $file->getThumbnail();
-                $data['thumbnail'] = !$thumbnail->isNew() ? $thumbnail->thumbnail : false;
-            }
-
-            $data['file'] = $file->toArray();
+        if (!$file->isNew()) {
+            $data['file']      = $file->toArray();
         }
 
         $data['created_on_timestamp']  = strtotime($this->created_on);
